Moved testing code back to 'code.rs'
